为面板数据增加圣遗物检测，防止错误数据
